backend/requirements.txt
django==4.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
slack-sdk==3.21.3